Bug 905 - Permission to use specific model elements should depend (also) on the model's domain
Added domain field in the permissions.xml file and related verification:
It is now possible to specify a domain in the permissions.xml file to
restrict usage of model elements to a particular domain. 
Syntax is: domain="<domainName>" 
Default is CROSS_DOMAIN (if no domain is specified, or the field is
missing, or has an invalid value). 

Change-Id: Id38d36800dbfa1374470770f9416908412a051fc
Signed-off-by: Laura Baracchi <laura.baracchi@intecs.it>